management_app_backend/requirements.txt
alembic==1.9.1
anyio==3.6.2
autopep8==2.0.1
click==8.1.3
ecdsa==0.18.0
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
Mako==1.2.4
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycodestyle==2.10.0
pydantic==1.10.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
tomli==2.0.1
typing-extensions==4.4.0
uvicorn==0.20.0
zipp==3.11.0
